添加多语言切换

+++ src/model/FileHandler.ts
@@ -3,6 +3,7 @@
  */
 
 import * as fs from 'fs';
+// import fs from 'fs/promises';
 import * as path from 'path';
 import { homedir } from "os";
 import { createHash } from 'crypto';
@@ -18,9 +19,9 @@ export class FileHandler {
      * 递归创建目录
      * @param dirPath 目录路径
      */
-    public static createDirectory(dirPath: string) {
+    public static async createDirectory(dirPath: string) {
         if (!fs.existsSync(dirPath)) {
-            this.createDirectory(path.dirname(dirPath));
+            await this.createDirectory(path.dirname(dirPath));
             fs.mkdirSync(dirPath);
         }
     }
@@ -29,8 +30,8 @@ export class FileHandler {
      * 创建文件
      * @param filePath 文件路径
      */
-    public static ensureDirectoryExistence(filePath: string, defaultValue: string = '') {
-        this.createDirectory(path.dirname(filePath));
+    public static async ensureDirectoryExistence(filePath: string, defaultValue: string = '') {
+        await this.createDirectory(path.dirname(filePath));
         if (!fs.existsSync(filePath)) {
             fs.writeFileSync(filePath, defaultValue);
         }
@@ -43,15 +44,25 @@ export class FileHandler {
      * @param dest 目标文件
      * @returns 
      */
-    public static copyFile(src: string, dest: string): boolean {
-        try {
-            this.createDirectory(path.dirname(dest));
-            fs.copyFileSync(src, dest);
-            return true
-        } catch (err) {
-            this.writeLog(err as string)
-            return false
-        }
+    public static copyFile(src: string, dest: string): Promise<boolean> {
+        return new Promise<boolean>(async (resolve, reject) => {
+            try {
+                await this.createDirectory(path.dirname(dest));
+                fs.copyFileSync(src, dest);
+                resolve(true)
+            } catch (err) {
+                this.writeLog(err as string)
+                reject(false)
+            }
+        })
+        // try {
+        //     this.createDirectory(path.dirname(dest));
+        //     fs.copyFileSync(src, dest);
+        //     return true
+        // } catch (err) {
+        //     this.writeLog(err as string)
+        //     return false
+        // }
     }
 
     /**
@@ -72,11 +83,11 @@ export class FileHandler {
                     const destFilePath = path.join(destPath, file);
                     // 如果是文件夹，则递归调用该文件夹下的所有文件和文件夹
                     if (fs.statSync(srcFilePath).isDirectory()) {
-                        this.createDirectory(destFilePath);
+                        await this.createDirectory(destFilePath);
                         await this.copyFolder(srcFilePath, destFilePath);
                     } else {
                         // 复制文件
-                        this.copyFile(srcFilePath, destFilePath)
+                        await this.copyFile(srcFilePath, destFilePath)
                     }
                 });
                 resolve(true)
@@ -137,8 +148,8 @@ export class FileHandler {
      * @param defaultValue 若文件不存在时的创建文件的默认值
      * @returns 
      */
-    public static readFile(filePath: string, defaultValue: string = ''): string {
-        this.ensureDirectoryExistence(filePath, defaultValue)
+    public static async readFile(filePath: string, defaultValue: string = ''): Promise<string> {
+        await this.ensureDirectoryExistence(filePath, defaultValue)
 
         let data = fs.readFileSync(filePath, 'utf-8')
         return data
@@ -154,6 +165,14 @@ export class FileHandler {
         return new Promise((resolve, reject) => {
             try {
                 this.createDirectory(path.dirname(filePath));
+
+                // 判断文件是否存在
+                if (!fs.existsSync(filePath)) {
+                    fs.writeFileSync(filePath, data);
+                    resolve(true)
+                    return
+                }
+
                 fs.writeFileSync(filePath, data);
                 resolve(true)
             } catch (err) {
@@ -223,15 +242,39 @@ export class FileHandler {
      * @param filePath 文件路径
      */
     public static openFile(filePath: string) {
+        console.log(filePath);
+
         exec(`open "${filePath}"`)
     }
 
     /**
      * 重命名文件
      */
-    public static renameFile(filePath: string, newName: string) {
-        fs.rename(filePath, newName, (err) => {
-            if (err) throw err;
+    public static renameFile(filePath: string, newName: string): Promise<void> {
+        return new Promise((resolve, reject) => {
+            fs.rename(filePath, newName, (err) => {
+                if (err) throw reject(err);
+                resolve()
+            })
         })
+    }
+
+    /**
+     * 获取文件大小
+     * @param filePath 文件路径
+     */
+    public static async getFileSize(filePath: string) {
+        const fileStat = await fs.promises.stat(filePath);
+        if (fileStat.isFile()) {
+            return fileStat.size;
+        } else if (fileStat.isDirectory()) {
+            let size = 0;
+            const files = await fs.promises.readdir(filePath);
+            for (let i = 0; i < files.length; i++) {
+                size += await this.getFileSize(path.join(filePath, files[i]));
+            }
+            return size;
+        }
+        return 0
     }
 }
